cleaning code

+++ backend/src/user/user.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, UseGuards, Req, Query, Post, Body, HttpCode, Param } from '@nestjs/common';
+import { Controller, Get, UseGuards, Req, Query, Post, Body, HttpCode, Param, HttpStatus } from '@nestjs/common';
 import { GetUser } from '../auth/decorator'
 import { FTAuthGuard, JwtGuard } from '../auth/guard';
 import { UserService } from './user.service';
@@ -23,6 +23,8 @@ export class UserController {
     }
     @Get('profile')
     async getProfile(@Req() req, @Query() query) {
+        if (!query.name)
+            throw HttpStatus.NOT_FOUND;
         return (await this.userService.getProfile(req.user.id, query.name));
     }
     @Get('list')
@@ -45,37 +47,51 @@ export class UserController {
     @Post('add/friend')
     @HttpCode(201)
     async addFriend(@Req() req, @Body() body) {
+        if (!body.name)
+            throw HttpStatus.BAD_REQUEST;
         await this.userService.addFriend(req.user.id, body.name);
     }
     @Post('accept/friend')
     @HttpCode(201)
     async acceptFriend(@Req() req, @Body() body) {
+        if (!body.name)
+            throw HttpStatus.BAD_REQUEST;
         return (await this.userService.acceptFriend(req.user.id, body.name));
     }
     @Post('block')
     @HttpCode(201)
     async blockUser(@Req() req, @Body() body) {
+        if (!body.name)
+            throw HttpStatus.BAD_REQUEST;
         return (await this.userService.blockUser(req.user.id, body.name));
     }
     @Post('inblock')
     @HttpCode(201)
     async inblockUser(@Req() req, @Body() body) {
+        if (!body.name)
+            throw HttpStatus.BAD_REQUEST;
         return (await this.userService.inblockUser(req.user.id, body.name));
     }
     @Post('remove/request')
     @HttpCode(201)
     async removeReq(@Req() req, @Body() body) {
+        if (!body.name)
+            throw HttpStatus.BAD_REQUEST;
         return (await this.userService.removeReq(req.user.id, body.name));
     }
     @Post('remove/friend')
     @HttpCode(201)
     async removeFriend(@Req() req, @Body() body) {
+        if (!body.name)
+            throw HttpStatus.BAD_REQUEST;
         return (await this.userService.removeFriend(req.user.id, body.name));    
     }
     //////ADD_ACHIEVEMENTS, GET_ACHIEVEMENTS///////
     @Post('achievements')
     @HttpCode(201)
     async addAchievement(@Req() req, @Body() body) {
+        if (body.achievement)
+            throw HttpStatus.BAD_REQUEST;
         return (await this.userService.addAchievement(req.user.id, body.achievement));
     }
     @Get('achievements')
